requirements.txt
langchain-openai
langchain
python-dotenv
pymongo
streamlit
fastapi
uvicorn
redis
